misc/Main-controller-app/lambda/requirements.txt
PyGithub==1.59.1
boto3==1.34.0
requests==2.31.0